maintenance: rebuildLocalisationCache should be DB_NONE if possible
This change should allow `rebuildLocalisationCache` to be run in a truly
offline environment as long as `LCStoreDB` is not being used.

Renames the `--no-clear-message-blob-store` option to `--offline` to
signal that no database dependent action will be performed.

Implements `getDbType()` which returns `DB_NONE` when `--offline` is
passed. This will avoid maintenance setup that tries to initialize a DB
load balancer.

Passing `--offline` when the store class is configured as `LCStoreDB`
will result in a fatal error.

Bug: T260827
Bug: T237148
Change-Id: I3e69b1ffad81a8c48f11a94973e7697461bb08ff
(cherry picked from commit d91c6627a9294d2a02526b4bb257178f51f8a782)