Do not lock user_preferences before updating
See T286521#7207425 for my theory on why deadlock
is happening. This is a stop-gap solution since the
ticket is a UBN, not a long-term fix.

This is also a partial revert of
Ibed2789f5260b725fd806b4470631aa30d814ce6.

The good: shouldn't deadlock.

The bad: it will be possible (but very unlikely) that
options save would undo previously saved options.

The long-term solution: we shouldn't need to re-fetch
all the options again before storing them. We now know
which options have changed. But the UserOptionsSave hook
requires all current options to be passed, and then the
hook can unset some of the options. Before implementing
the long-term solution, we need to change the hook
to only need modified options, not all the options.
This will be done as a followup.

Bug: T286521
Change-Id: Ie6203be29ab23bc7ef9e26db421365a4f075c78f
(cherry picked from commit 98bb3df8162c21cd84cada872ed380116c37d88e)