resourceloader: Fix prepending of OOUI theme skinStyles
== Background

Prior to 1367e356e76e:

* Module definition is registered (may have skinStyles).
* If the module already has skinStyles, skip this step.
  Otherwise, skin-provided ResourceModuleSkinStyles are merged
  into the module def, where '+module' means defaults are kept,
  and 'module' replaces the defaults.
* When the module is constructed for bundling,
  OOUIFileModule prepends theme-specific resources to skinStyles.

For OOUI and MonoBook that meant:

* oojs-ui-core.styles has no own skinStyles.
* MonoBook/skin.json defines additive (+module) styles,
  which are set into an otherwise empty skinStyles array.
* OOUIFileModule sees skinStyles[monobook] and prepends to it.

After 1367e356e76e, it works like this:

* Module definition is registered.
* The module is constructed, and OOUIFileModule sets
  skinStyles[monobook] as new key.
* Skin-provided ResourceModuleSkinStyles is ignored because by design
  these are not allowed if the module has skinStyles[monobook] set
  itself.

This refactor changed nothing for most module classes, but did for
OOUIFileModule.

== Fix

We can move the OOUIFileModule logic up to where the new abstraction
is, or we can move it down toward run-time methods like getStyleFiles
(as SkinModule currently does).

I'm choosing the former as otherwise we lose automatic versioning,
caching, and inclusion in PHPUnit structure tests for file existence
and Less syntax etc.

Bug: T290013
Change-Id: Id7c258841d7816e710961a1f7c7f7c99764cf59a