shell: Fix $wgShellboxUrls by passing service name when creating BoxedCommand
To be able to use non-default Shellbox URLs, we need to be able to
pass in a service name when creating the ShellboxClient.

Have CommandFactory::createBoxed() take a $service parameter that can
be used to change which Shellbox will be hit, as intended.

Bug: T290193
Change-Id: Ic1671a69070f962dbb0083028faf34d6d437022a
(cherry picked from commit b6d1afa1aa851874d1e818ec9781d833d24b8330)