uppercaseTitlesForUnicodeTransition: improve userlist format
This script is used to rename things to prepare for unicode
update. The script does not rename the users, instead it generates
a list of users to be renamed with renameInvalidUsernames.php,
but this script output is incompatible with the other script input.
It would be more convinient to be able to directly feed the result
of this script as input to the other script.

renameInvalidUsernames.php expects input as TSV with domain, user_id,
new_user_name. That's what we want to output. Also opening the
output file in append mode is more convenient - that way we can
collect all usernames for renames for all wikis when running
the script with foreachwiki

Bug: T219279
Change-Id: I27e182247d55c635e75027065546285be74934b6
(cherry picked from commit 756eefa6bf7e7c79515f7b1d8803d55a375d11d5)