MultiHttpClient: Allow setting HTTP protocol version in curl
For context see T275752, but the premise is that we want to be able to
set the HTTP protocol version so that for specific use cases. That way
we will be able to use the proper protocol per use case, mitigating
potential future regressions or selectively rolling out a newer protocol
version per use case. Support setting CURLOPT_HTTP_VERSION, while making
sure to default to the current default which lets curl decide which
version to use

CurlHttpRequest already forcibly set this to 1.0 for what is worth

Bug: T275752
Change-Id: I82f4f174997ecaa54e33eb848b5f007c982506dd
(cherry picked from commit afe2d229c9776387017a6ce34b115b879bf7aff1)