Force using HTTP 1.1 for SwiftFileBackend
We want to force the use of HTTP/1.1 instead of leaving it up to the host
default which might be HTTP/2, introducing a performance regression for
large uploads.

Since we have to pass this to every run()/runMulti(), set it as a private
constant that can be passed directlry or added to existing option arrays.

Bug: T275752
Change-Id: Ifbae4d585daee28fa92b07f53c5854320a18a311
(cherry picked from commit 5a62ea7d7747f20c6e8d990fa83dc15f70c68ee9)