First pass on creating config-schema.yaml
The config-schema.yaml follows the settings file format.
It's not used in reality yet, but tests were added
to compare the yaml with the result of DefaultSettings.php
importing.

The list of properties not yet included:
 - 'wgVersion' - deprecated alias for MW_VERSION
 - 'wgConf' - instance of SiteConfiguration
 - 'wgStyleSheetPath' - alias to another global
 - 'wgExtensionDirectory', - depends on $IP
 - 'wgStyleDirectory' - depends on $IP
 - 'wgServiceWiringFiles' - depends on __DIR__
 - 'wgHTTPMaxTimeout' - infinity default
 - 'wgHTTPMaxConnectTimeout', - Infinity default

The PHPDoc was copied over as is into description fields,
it will be converted to something more pleasant in a separate
iteration.

The types were parsed out of PHPDoc and fixed up manually.

Default values come from executing DefaultSettings and serializing
the values into YAML.

Change-Id: I9aa9ad9713bdbdac2bd7b528c609772154b5a135