Reuse the query result in addCategoryLinks instead of relying on cache
This is an implicit logic that it first loads everything as a batch and
then adds them to the LinkCache and then just hopes it works.

During passing arround of Title object, it reaches Title::getArticleID()
(in Title::exists()) and that now reads it from PageStore
(by using ::getFieldFromPageStore) and since the Page object doesn't
exist yet (we only populated the LinkCache), it loads the page again.

Bug: T297669
Change-Id: Icca81c97a15d90f593090711cdccdee3ea404ae2
(cherry picked from commit 70acd1dcd40d8fe573ac9074c133e003142eac81)