In WikitextContentHandler always use getFreshParser()
Make it safe to parse articles while in the parser, by always calling
getFreshParser() from WikitextContentHandler.

I think ideally this should be a ParserFactory responsibility, with
Parser instances stored by ParserFactory instead of directly by
ServiceContainer, but this fixes the bug, follows existing conventions,
and does not reduce performance in the usual case.

Bug: T299149
Change-Id: Ibbc1423166f4804a5122de10293ea26f5704d96d